move response types to own file
