+++ Cargo.toml
@@ -10,5 +10,4 @@ dotenv = "0.15.0"
 chrono = { version ="0.4.23", features = ["serde", "rustc-serialize"] }
 reqwest = { version = "0.11.13", features = ["json"] }
 serde = { version = "1.0.151", features=["derive"] }
-serde_json = "1.0"
 tokio = { version = "1.23.0", features = ["full"] }
